Brand 회원가입 분리부분
- verification 데이터의 삭제 플레그값 추가

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/service/AuthService.java
@@ -130,11 +130,13 @@ public class AuthService {
 	 * @return 존재여부
 	 */
 	@Transactional(rollbackFor = Exception.class)
-	public boolean checkVerificationCode(String key, String code) {
+	public boolean checkVerificationCode(String key, String code, boolean boolDelete) {
 		if (!repoVerificationCode.findByVerificationKeyAndCode(key, code).isPresent()) {
 			return false;
 		}
-		repoVerificationCode.deleteById(key);
+		if(boolDelete){
+			repoVerificationCode.deleteById(key);
+		}
 		return true;
 	}
